refactor: extract community follows hook
